hawkbit-client: minimize sleep time during processing/downloading/cancel requested
All "requests" from hawkBit are part of the base resource. In order to
react in a timely fashion to those requests when another operation is in
progress (processing/downloading/canceling), we need to minimize the
polling interval during those operations. On the other hand, we do not
want to flood hawkBit (imagine a huge rollout and all targets poll
constantly for updates while processing/downloading the update).

Polling every 5 seconds during these operations is implemented as a
trade-off.

Signed-off-by: Bastian Krause <bst@pengutronix.de>